+++ ui/package.json
@@ -1,39 +1,40 @@
 {
-	"name":"storanizer-ui",
-	"version":"1.0.0",
-	"description":"",
-	"main":"index.js",
+	"name": "storanizer-ui",
+	"version": "1.0.0",
+	"description": "",
+	"main": "index.js",
 	"scripts": {
-		"test":"nyc --require esm --require jsdom-global/register --require @riotjs/register mocha src/**/*.spec.js",
-		"cov":"nyc report --reporter=text-lcov",
-		"cov-html":"nyc report --reporter=html",
-		"build":"npx webpack --mode production",
-		"prepublishOnly":"npm test",
-		"start":"npx webpack serve --mode development --hot --port 3000"
+		"test": "nyc --require esm --require jsdom-global/register --require @riotjs/register mocha src/**/*.spec.js",
+		"cov": "nyc report --reporter=text-lcov",
+		"cov-html": "nyc report --reporter=html",
+		"build": "npx webpack --mode production",
+		"prepublishOnly": "npm test",
+		"start": "npx webpack serve --mode development --hot --port 3000"
 	},
-	"author":"",
-	"license":"ISC",
+	"author": "",
+	"license": "ISC",
 	"devDependencies": {
-		"@riotjs/compiler":"^6.3.2",
-		"@riotjs/register":"^6.0.1",
-		"@riotjs/webpack-loader":"^6.0.0",
-		"chai":"^4.3.6",
-		"css-loader":"^6.7.1",
-		"esm":"^3.2.25",
-		"html-webpack-plugin":"^5.5.0",
-		"jsdom":"^20.0.0",
-		"jsdom-global":"^3.0.2",
-		"mini-css-extract-plugin":"^2.6.1",
-		"mocha":"^8.4.0",
-		"nyc":"^15.1.0",
-		"webpack":"^5.74.0",
-		"webpack-cli":"^4.10.0",
-		"webpack-dev-server":"^4.10.0"
+		"@riotjs/compiler": "^6.3.2",
+		"@riotjs/register": "^6.0.1",
+		"@riotjs/webpack-loader": "^6.0.0",
+		"chai": "^4.3.6",
+		"css-loader": "^6.7.1",
+		"esm": "^3.2.25",
+		"html-webpack-plugin": "^5.5.0",
+		"jsdom": "^20.0.0",
+		"jsdom-global": "^3.0.2",
+		"mini-css-extract-plugin": "^2.6.1",
+		"mocha": "^8.4.0",
+		"nyc": "^15.1.0",
+		"webpack": "^5.74.0",
+		"webpack-cli": "^4.10.0",
+		"webpack-dev-server": "^4.10.0"
 	},
 	"dependencies": {
-		"@riotjs/hot-reload":"^6.0.1",
-	 	"@riotjs/lazy":"^2.0.2",
-	 	"@riotjs/route":"^8.0.2",
-	 	"riot":"^7.0.3"
-	 }
+		"@riotjs/custom-elements": "^6.0.1",
+		"@riotjs/hot-reload": "^6.0.1",
+		"@riotjs/lazy": "^2.0.2",
+		"@riotjs/route": "^8.0.2",
+		"riot": "^7.0.3"
+	}
 }
