+++ ui/src/components/controller/items/item.riot
@@ -1,8 +1,8 @@
 <item>
     <h2>🛠️ { state.item.name }</h2>
     <p>{ state.item.description }</p>
-    <p if={ state.box.id }><b>Box: </b><a href="#box/{ state.box.id }">{ state.box.name }</a></p>
-    <p if={ state.location.id }><b>Location: </b><a href="#location/{ state.location.id }">{ state.location.name }</a></p>
+    <p if={ state.box.id }><b>Box: </b><a onclick={ window.application.clickHandler } href="#box/{ state.box.id }">{ state.box.name }</a></p>
+    <p if={ state.location.id }><b>Location: </b><a onclick={ window.application.clickHandler } href="#location/{ state.location.id }">{ state.location.name }</a></p>
     <qr-code></qr-code>
   <script>
     import  './style.css'
