+++ ui/src/components/controller/locations/location.riot
@@ -44,6 +44,9 @@
       },
 
       onMounted() {
+        console.log('now or never')
+        console.log(this)
+        console.log(window)
         this.loadData()
       }
     }
