+++ ui/src/components/controller/locations/locations.riot
