--- ui/src/components/global/login/login.riot
@@ -1,55 +1,0 @@
-<login>
-  <div class="row">
-    <div if={ !state.login } class="hello column column-67">
-      <h1><img width="35px" src="./storganizer-logo.png" /> Storganizer</h1>
-      <p>Hello There</p>
-    </div>
-
-    <div class="login column column-33">
-      <div class="row">
-        <h1>Login</h1>
-      </div>
-
-      <div if={state.showUser} class="row">
-        <user name={ state.name }/>
-      </div>
-
-      <div class="row username">
-        <input type="text" placeholder="Username" id="username" name="username">
-      </div>
-
-      <div class="row password">
-        <input type="password" placeholder="only god and I knows"  id="password" name="password">
-      </div>
-
-      <div class="row">
-        <button onclick={toggleUser}>login</button>
-      </div>
-    </div>
-  </div>
-
-  <script>
-    import User from '../../includes/user/user.riot'
-    import  './style.css'
-
-    export default {
-      components: {
-        User
-      },
-      state: {
-        name: 'Claudio',
-        showUser: false
-      },
-      toggleUser() {
-        this.update({
-          showUser: !this.state.showUser
-        })
-      },
-      login() {
-        return true
-      }
-    }
-  </script>
-
-</login>
-
