--- ui/src/components/global/login/style.css
@@ -1,20 +1,0 @@
-login {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  min-height: 100vh;
-}
-
-.hello {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.hello p {
-    text-align: center;
-}
-
-.login {
-  min-width: 300px;
-}
