--- ui/src/components/includes/user/user.riot
@@ -1,3 +1,0 @@
-<user>
-  <p>My  username is: <b>{ props.name }</b></p>
-</user>
